requirements.txt
anthropic>=0.28.0
tavily-python>=0.3.0
gspread>=6.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
matplotlib>=3.8.0
Pillow>=10.3.0
Jinja2>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
html2text>=2024.2.26
